Move Path utility into ContentConfig
Resolves #218
